MD-989 sentinel: try next sentinels after I/O timeout errors
This fixes an issue that causes other sentinels not to be tried after
one of them fails with an I/O timeout.

The root cause is that the code attempts to abort early if the error
is of one of the types generated by the context module, in this case,
`context.DeadlineExceeded`, but it doesn't check that the error comes
from the context argument provided to the function. Since the network
layer also uses context to manage timeouts, it returns the same type
of error, which is mistakenly assumed as coming from the passed
context.

The fix consists of checking whether the `ctx` object passed to the
function is in error state to abort early, and continue trying the
next sentinels otherwise.

The commit that introduced the early abortion on context error and
caused this regression is:

https://github.com/redis/go-redis/commit/63392a363ad7dce4c1b5aae60d1207104664371b